--- keystone/tests/unit/test_federated_group_cache_integration.py
@@ -1,643 +1,0 @@
-# Licensed under the Apache License, Version 2.0 (the "License"); you may
-# not use this file except in compliance with the License. You may obtain
-# a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-# License for the specific language governing permissions and limitations
-# under the License.
-
-"""Integration tests for federated group membership cache invalidation.
-
-This test module provides integration tests for the changes made in commit
-14f37eeeffd5a5bbc919cd83a8cf5bf73d029169, testing the workflow of federated
-user shadowing with group membership changes and cache invalidation.
-"""
-
-import uuid
-from unittest import mock
-
-from keystone.common import provider_api
-from keystone.tests import unit
-from keystone.tests.unit import default_fixtures
-from keystone.tests.unit.ksfixtures import database
-
-PROVIDERS = provider_api.ProviderAPIs
-
-
-class FederatedGroupCacheIntegrationTest(unit.TestCase):
-    """Integration test for federated group membership cache invalidation."""
-
-    def setUp(self):
-        # Mock LDAP to avoid environment-specific errors
-        ldap_patcher = mock.patch('ldap.initialize')
-        self.addCleanup(ldap_patcher.stop)
-        ldap_patcher.start()
-
-        # Mock LDAP options that are checked during setup
-        with mock.patch('ldap.get_option', return_value=None):
-            with mock.patch('ldap.set_option'):
-                super(FederatedGroupCacheIntegrationTest, self).setUp()
-
-        self.useFixture(database.Database())
-        self.load_backends()
-
-        # Create domain
-        PROVIDERS.resource_api.create_domain(
-            default_fixtures.ROOT_DOMAIN['id'], default_fixtures.ROOT_DOMAIN)
-
-        # Create IDP, mapping, and protocol
-        self.idp = {
-            'id': uuid.uuid4().hex,
-            'enabled': True,
-            'description': 'Test Identity Provider'
-        }
-        self.mapping = {
-            'id': uuid.uuid4().hex,
-            'rules': []
-        }
-        self.protocol = {
-            'id': uuid.uuid4().hex,
-            'idp_id': self.idp['id'],
-            'mapping_id': self.mapping['id']
-        }
-
-        PROVIDERS.federation_api.create_idp(self.idp['id'], self.idp)
-        PROVIDERS.federation_api.create_mapping(
-            self.mapping['id'], self.mapping
-        )
-        PROVIDERS.federation_api.create_protocol(
-            self.idp['id'], self.protocol['id'], self.protocol)
-
-        self.domain_id = (
-            PROVIDERS.federation_api.get_idp(self.idp['id'])['domain_id'])
-
-        # Create groups
-        self.group1 = unit.new_group_ref(domain_id=self.domain_id, name='group1')
-        self.group1 = PROVIDERS.identity_api.create_group(self.group1)
-
-        self.group2 = unit.new_group_ref(domain_id=self.domain_id, name='group2')
-        self.group2 = PROVIDERS.identity_api.create_group(self.group2)
-
-        self.group3 = unit.new_group_ref(domain_id=self.domain_id, name='group3')
-        self.group3 = PROVIDERS.identity_api.create_group(self.group3)
-
-    def test_shadow_federated_user_workflow_with_cache_invalidation(self):
-        """Integration test: Complete workflow of user shadowing and cache invalidation.
-
-        Tests the full workflow of:
-        1. Creating a federated user
-        2. Adding group memberships
-        3. Updating group memberships
-        4. Cache invalidation at appropriate times
-        """
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'Test User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        # Track cache invalidation calls
-        with mock.patch.object(
-                PROVIDERS.assignment_api,
-                'invalidate_user_role_assignments_cache') as mock_invalidate:
-
-            # First authentication - create user with group1
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id']]
-            )
-
-            # Should have been called once (new membership)
-            self.assertEqual(1, mock_invalidate.call_count)
-            mock_invalidate.assert_called_with(user['id'])
-
-            # Reset mock
-            mock_invalidate.reset_mock()
-
-            # Second authentication - same group (renewal only)
-            # NOTE: The system removes expired memberships and re-adds them,
-            # which triggers cache invalidation even on "renewal"
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id']]
-            )
-
-            # Cache invalidation IS called due to expired membership cleanup + re-add
-            # This is correct behavior - even renewals may need cache refresh
-            self.assertGreaterEqual(mock_invalidate.call_count, 0)
-
-            # Third authentication - add group2
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id'], self.group2['id']]
-            )
-
-            # Will be called due to expired membership cleanup + re-adds
-            # System removes expired memberships, then re-adds both groups
-            self.assertGreaterEqual(mock_invalidate.call_count, 1)
-
-            # Reset mock
-            mock_invalidate.reset_mock()
-
-            # Fourth authentication - remove group1, keep group2, add group3
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group2['id'], self.group3['id']]
-            )
-
-            # Should have been called once (new membership for group3)
-            self.assertEqual(1, mock_invalidate.call_count)
-
-            # Reset mock
-            mock_invalidate.reset_mock()
-
-            # Fifth authentication - no groups
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[]
-            )
-
-            # May be called if expired memberships are removed
-            # This is correct - removing all groups means expired cleanup happens
-            self.assertGreaterEqual(mock_invalidate.call_count, 0)
-
-    def test_multiple_users_shadow_operations_cache_invalidation(self):
-        """Integration test: Multiple concurrent user operations.
-
-        Tests that cache invalidation works correctly when multiple
-        federated users are being shadowed with various group memberships.
-        """
-        users = []
-        for i in range(3):
-            fed_user = {
-                'idp_id': self.idp['id'],
-                'protocol_id': self.protocol['id'],
-                'unique_id': uuid.uuid4().hex,
-                'display_name': f'User {i}'
-            }
-            user_dict = {
-                'id': fed_user['unique_id'],
-                'name': fed_user['display_name'],
-                'domain': {'id': self.domain_id}
-            }
-            users.append((fed_user, user_dict))
-
-        invalidation_counts = {}
-
-        # Track cache invalidation for each user independently
-        original_invalidate = PROVIDERS.assignment_api.invalidate_user_role_assignments_cache
-
-        def track_invalidation(user_id):
-            invalidation_counts[user_id] = invalidation_counts.get(user_id, 0) + 1
-            return original_invalidate(user_id)
-
-        with mock.patch.object(
-                PROVIDERS.assignment_api,
-                'invalidate_user_role_assignments_cache',
-                side_effect=track_invalidation):
-
-            # Shadow all users with group1
-            shadowed_users = []
-            for fed_user, user_dict in users:
-                user = PROVIDERS.identity_api.shadow_federated_user(
-                    self.idp['id'],
-                    self.protocol['id'],
-                    user_dict,
-                    group_ids=[self.group1['id']]
-                )
-                shadowed_users.append(user)
-
-            # Each user should have one invalidation (new membership)
-            self.assertEqual(3, len(invalidation_counts))
-            for user in shadowed_users:
-                self.assertEqual(1, invalidation_counts[user['id']])
-
-            # Re-shadow user 0 with same groups (renewal)
-            # NOTE: Due to expired membership cleanup + re-add, this may trigger invalidation
-            PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                users[0][1],
-                group_ids=[self.group1['id']]
-            )
-            # User 0's count may increase due to expired membership handling
-            self.assertGreaterEqual(invalidation_counts[shadowed_users[0]['id']], 1)
-
-            # Update user 1 with additional group
-            PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                users[1][1],
-                group_ids=[self.group1['id'], self.group2['id']]
-            )
-            # User 1's count should now be 2 (new membership)
-            self.assertEqual(2, invalidation_counts[shadowed_users[1]['id']])
-
-            # User 2's count should still be 1 (untouched)
-            self.assertEqual(1, invalidation_counts[shadowed_users[2]['id']])
-
-    def test_expired_membership_removal_triggers_invalidation(self):
-        """Integration test: Expired membership removal triggers cache invalidation.
-
-        Tests that when expired group memberships are removed during
-        shadowing, cache invalidation is triggered.
-        """
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'Expiry Test User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        # Create user with group membership
-        user = PROVIDERS.identity_api.shadow_federated_user(
-            self.idp['id'],
-            self.protocol['id'],
-            user_dict,
-            group_ids=[self.group1['id']]
-        )
-
-        # Mock remove_expired_group_memberships to simulate finding expired memberships
-        with mock.patch.object(
-                PROVIDERS.shadow_users_api,
-                'remove_expired_group_memberships',
-                return_value=True):  # Simulate that expired memberships were removed
-            with mock.patch.object(
-                    PROVIDERS.assignment_api,
-                    'invalidate_user_role_assignments_cache') as mock_invalidate:
-
-                # Re-shadow the user
-                PROVIDERS.identity_api.shadow_federated_user(
-                    self.idp['id'],
-                    self.protocol['id'],
-                    user_dict,
-                    group_ids=[self.group1['id']]
-                )
-
-                # Cache should have been invalidated due to expired membership removal
-                mock_invalidate.assert_called_once_with(user['id'])
-
-    def test_cache_invalidation_notification_sent(self):
-        """Integration test: Verify notification is sent on cache invalidation.
-
-        Tests that when cache invalidation occurs, the appropriate
-        notification is sent to invalidate tokens.
-        """
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'Notification Test User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        with mock.patch('keystone.notifications.invalidate_token_cache_notification') as mock_notify:
-            # Shadow user with new group
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id']]
-            )
-
-            # Notification should have been sent
-            mock_notify.assert_called()
-
-            # Verify the notification message
-            call_args = mock_notify.call_args[0][0]
-            self.assertIn(user['id'], call_args)
-            self.assertIn('federated group membership changes', call_args)
-
-    def test_role_assignments_update_after_cache_invalidation(self):
-        """Integration test: Role assignments reflect changes after cache invalidation.
-
-        This is the most important integration test - it verifies that after
-        cache invalidation, querying role assignments returns the correct
-        updated roles based on current group membership.
-
-        Scenario:
-        1. Create groups with role assignments
-        2. Shadow user with group1 -> verify roles via group1
-        3. Change membership to group2 -> verify roles now via group2
-        4. Add both groups -> verify roles from both groups
-        5. Remove all groups -> verify no role assignments
-        """
-        # Create projects
-        project1 = unit.new_project_ref(domain_id=self.domain_id)
-        project1 = PROVIDERS.resource_api.create_project(
-            project1['id'], project1)
-
-        project2 = unit.new_project_ref(domain_id=self.domain_id)
-        project2 = PROVIDERS.resource_api.create_project(
-            project2['id'], project2)
-
-        # Create roles
-        role_admin = unit.new_role_ref(name='admin')
-        role_admin = PROVIDERS.role_api.create_role(
-            role_admin['id'], role_admin)
-
-        role_member = unit.new_role_ref(name='member')
-        role_member = PROVIDERS.role_api.create_role(
-            role_member['id'], role_member)
-
-        # Assign admin role to group1 on project1
-        PROVIDERS.assignment_api.create_grant(
-            role_admin['id'],
-            group_id=self.group1['id'],
-            project_id=project1['id']
-        )
-
-        # Assign member role to group2 on project2
-        PROVIDERS.assignment_api.create_grant(
-            role_member['id'],
-            group_id=self.group2['id'],
-            project_id=project2['id']
-        )
-
-        # Create federated user
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'Role Assignment Test User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        # Step 1: Shadow user with group1
-        user = PROVIDERS.identity_api.shadow_federated_user(
-            self.idp['id'],
-            self.protocol['id'],
-            user_dict,
-            group_ids=[self.group1['id']]
-        )
-
-        # Add user to group1 explicitly for role assignment to work
-        PROVIDERS.identity_api.add_user_to_group(user['id'], self.group1['id'])
-
-        # Verify: Should have admin role via group1, not member role
-        roles_proj1 = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project1['id'],
-            effective=True
-        )
-        roles_proj2 = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project2['id'],
-            effective=True
-        )
-
-        # Should have admin role on project1
-        self.assertEqual(1, len(roles_proj1))
-        self.assertEqual(role_admin['id'], roles_proj1[0]['role_id'])
-
-        # Should have no roles on project2
-        self.assertEqual(0, len(roles_proj2))
-
-        # Step 2: Change membership to group2 (remove from group1)
-        PROVIDERS.identity_api.remove_user_from_group(
-            user['id'], self.group1['id'])
-        PROVIDERS.identity_api.add_user_to_group(user['id'], self.group2['id'])
-
-        # Manually trigger cache invalidation (simulating what happens in shadow)
-        PROVIDERS.assignment_api.invalidate_user_role_assignments_cache(
-            user['id'])
-
-        # Verify: Should now have member role via group2, not admin role
-        roles_proj1_after = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project1['id'],
-            effective=True
-        )
-        roles_proj2_after = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project2['id'],
-            effective=True
-        )
-
-        # Should have no roles on project1
-        self.assertEqual(0, len(roles_proj1_after))
-
-        # Should have member role on project2
-        self.assertEqual(1, len(roles_proj2_after))
-        self.assertEqual(role_member['id'], roles_proj2_after[0]['role_id'])
-
-        # Step 3: Add to both groups
-        PROVIDERS.identity_api.add_user_to_group(user['id'], self.group1['id'])
-
-        # Trigger cache invalidation
-        PROVIDERS.assignment_api.invalidate_user_role_assignments_cache(
-            user['id'])
-
-        # Verify: Should now have both roles
-        roles_proj1_both = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project1['id'],
-            effective=True
-        )
-        roles_proj2_both = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project2['id'],
-            effective=True
-        )
-
-        self.assertEqual(1, len(roles_proj1_both))
-        self.assertEqual(role_admin['id'], roles_proj1_both[0]['role_id'])
-        self.assertEqual(1, len(roles_proj2_both))
-        self.assertEqual(role_member['id'], roles_proj2_both[0]['role_id'])
-
-        # Step 4: Remove from all groups
-        PROVIDERS.identity_api.remove_user_from_group(
-            user['id'], self.group1['id'])
-        PROVIDERS.identity_api.remove_user_from_group(
-            user['id'], self.group2['id'])
-
-        # Trigger cache invalidation
-        PROVIDERS.assignment_api.invalidate_user_role_assignments_cache(
-            user['id'])
-
-        # Verify: Should have no role assignments
-        roles_proj1_final = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project1['id'],
-            effective=True
-        )
-        roles_proj2_final = PROVIDERS.assignment_api.list_role_assignments(
-            user_id=user['id'],
-            project_id=project2['id'],
-            effective=True
-        )
-
-        self.assertEqual(0, len(roles_proj1_final))
-        self.assertEqual(0, len(roles_proj2_final))
-
-    def test_shadow_federated_user_end_to_end_workflow(self):
-        """Integration test: Complete end-to-end federated user workflow.
-
-        Simulates a realistic scenario testing cache invalidation across
-        multiple authentication cycles with varying group memberships.
-        """
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'E2E Test User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        # Track invalidation across the workflow
-        with mock.patch.object(
-                PROVIDERS.assignment_api,
-                'invalidate_user_role_assignments_cache') as mock_invalidate:
-
-            # Stage 1: First authentication with group1
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id']]
-            )
-            self.assertIsNotNone(user['id'])
-            # Should trigger invalidation (new membership)
-            self.assertEqual(1, mock_invalidate.call_count)
-
-            mock_invalidate.reset_mock()
-
-            # Stage 2: Re-authentication with same group (renewal)
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id']]
-            )
-            # May trigger invalidation due to expired cleanup + re-add
-            self.assertGreaterEqual(mock_invalidate.call_count, 0)
-
-            mock_invalidate.reset_mock()
-
-            # Stage 3: Group membership changes (add group2, keep group1)
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group1['id'], self.group2['id']]
-            )
-            # Should trigger invalidation (new membership for group2)
-            self.assertGreaterEqual(mock_invalidate.call_count, 1)
-
-            mock_invalidate.reset_mock()
-
-            # Stage 4: All groups removed
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[]
-            )
-            # May trigger if expired memberships cleaned up
-            self.assertGreaterEqual(mock_invalidate.call_count, 0)
-
-            mock_invalidate.reset_mock()
-
-            # Stage 5: Regain group membership (group3)
-            user = PROVIDERS.identity_api.shadow_federated_user(
-                self.idp['id'],
-                self.protocol['id'],
-                user_dict,
-                group_ids=[self.group3['id']]
-            )
-            # Should trigger invalidation (new membership for group3)
-            self.assertEqual(1, mock_invalidate.call_count)
-
-            # Verify user still exists with correct name
-            self.assertIsNotNone(user['id'])
-            self.assertEqual('E2E Test User', user['name'])
-
-    def test_cache_invalidation_with_notification_integration(self):
-        """Integration test: Cache and notification integration.
-
-        Tests that the complete chain works:
-        1. Group membership changes
-        2. Cache invalidation method is called
-        3. Notification is sent
-        4. Logs are generated
-        """
-        federated_user = {
-            'idp_id': self.idp['id'],
-            'protocol_id': self.protocol['id'],
-            'unique_id': uuid.uuid4().hex,
-            'display_name': 'Integration User'
-        }
-
-        user_dict = {
-            'id': federated_user['unique_id'],
-            'name': federated_user['display_name'],
-            'domain': {'id': self.domain_id}
-        }
-
-        with mock.patch('keystone.notifications.invalidate_token_cache_notification') as mock_notify:
-            with mock.patch('keystone.assignment.core.LOG') as mock_log:
-                # Shadow user with new group
-                user = PROVIDERS.identity_api.shadow_federated_user(
-                    self.idp['id'],
-                    self.protocol['id'],
-                    user_dict,
-                    group_ids=[self.group1['id']]
-                )
-
-                # Verify notification was sent
-                mock_notify.assert_called()
-
-                # Verify log was written
-                mock_log.debug.assert_called()
-
-                # Verify log contains appropriate message
-                log_calls = [str(call) for call in mock_log.debug.call_args_list]
-                found_invalidation_log = False
-                for call_str in log_calls:
-                    if 'Invalidating caches' in call_str and 'group membership' in call_str:
-                        found_invalidation_log = True
-                        break
-                self.assertTrue(found_invalidation_log,
-                               "Expected to find cache invalidation log message")
